pyproject.toml
[tool.poetry]
name = "kaggle-data-analysis"
version = "0.1.0"
description = ""
authors = ["uowol <alllhvhla@gmail.com>"]
readme = "README.md"


[tool.black]
line-length = 88
target-version = ['py312']
skip-string-normalization = false


[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
known_third_party = ["numpy", "pandas", "seaborn", "scikit-learn"]


[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = "^2.2.3"
numpy = "2.1.0"
matplotlib = "^3.10.0"
seaborn = "^0.13.2"
requests = "^2.32.3"
ipykernel = "^6.29.5"
black = "^24.10.0"
isort = "^5.13.2"
kaggle = "^1.6.17"
setuptools = "^75.8.0"
pyaml = "^25.1.0"
pydantic = "^2.10.6"
ydata-profiling = "^4.12.2"
statsmodels = "^0.14.4"
scikit-learn = "^1.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
